+++ package.js
@@ -9,9 +9,8 @@ Npm.depends({
 Package.on_use(function(api) {
 	// meteor-headers 0.0.7 was the last version to support Meteor < 0.6.5
 	api.use(['webapp', 'livedata', 'templating', 'deps'], ['client', 'server']);
-	api.use(['appcache', 'browser-policy'], 'server', { weak: true });
-
-	api.use('standard-app-packages', 'client');
+	api.use('appcache', 'server', { weak: true });
+	api.use('inject-initial', ['server', 'client']);
 
     api.add_files('headers-common.js', ['client', 'server']);
     api.add_files('headers-server.js', 'server');